feat: Add course of study to user profile and registration
Introduces the 'courseOfStudy' field to the User type and registration data. This allows users to specify their course of study during registration, and this information is now stored in their profile.

Updates the API service to handle the new field during user creation and the StudentAuth page to include it in the registration form. Also, minor UI adjustments were made to the welcome carousel.